Update Readme regarding fiat currencies
Since the Readme was written, there are more currencies available
for the users. They are listed in
`public/mbwapi/src/main/java/com/mrd/mbwapi/api/CurrencyCode.java`,
and currently I could 164 of them.

Update readme to be more up to date, though approximate number is enough,
does not need to mention the exact number of supported currencies,
otherwise it's a maintenance nightmare whenever a new currency is added.