unit tests for ColuAssetUri found a bug in toString()
there was a hard coded RMC in toString of ColuAssetUri
other tests are failing according to how I understand that it should work but maybe I'm wrong. My understanding is that the Mycelium wallet should not under any circumstance interpret a bitcoin:... uri as a ColuAssetUri of the **scheme bitcoin**. Also would I assume the scheme to be case insensitive. Also does ColuAccount.ColuAssetType imply that mass tokens can either have the scheme mass or mss.