testnet wallet hack
due to our brief featuring of our tech demo on the testnet wallet application id, that app can now only be uploaded to google play targetting api 23 and above.
api 23 means dynamic permissions
our support libraries are stuck at api 19 (and need to be updated, too)
the fastest hack to only help people with crashing testnet wallets was to toast the user what to do in the app settings and send him to the app settings. This change should only affect testnet.