minor changes and killed some 4 years old migration code
if somebody did not update his wallet in 4 years, he loses address
labels in exchange for a slightly faster startup for all. In general we
need to put version updates into a version manager that always runs only
the necessary changes instead of doing all tests all the time for all
errors we fixed in the past.