fix #728 use credentials according to currency pair
Mycelium uses different accounts for BCH2BTC than for other pairs.
This patch is a bit hacky as it defaults to the "other pairs"
credentials for calls that don't involve a currency pair (getCurrencies
for example).
In order to implement it more correct, the ChangellyService should
probably get refactored out. I'm convinced it is an unnecessary
indirection. With it gone, the BCH2BTC assistant could inject one set of
credentials and the receive-from-alts assistant could inject another.