check for version when pairing
we regularly break compatibility between mbw and the modules, when we
upgrade the contract or how it is interpreted.
with this change, mbw and the module hold expections on the
modularization protocol version and pairing fails if it is not met.
TODO: Tell the user to upgrade the module
TODO: avoid crashing