turned AsyncTask into static class with WeakReferences
lint warnings about
This AsyncTask class should be static or leaks might occur
was bothering me since a long time. here I tried to make it right, with
static class and WeakReferences.
As this AsyncTask is particularly big, the fix was also a bit more
complicated than it would be for most other AsyncTasks we have, with
custom Exception to early-out from nested method calls for example.