new interface ServerListChangedListener
WalletConfiguration shouldn't get the whole MbwManager just to get
notified of a new configuration.
Also the configuration needs updating if the new endpoints are different
to the old ones, not to the default ones, so the default ones were
removed, too. Lastly, the endpoint changes only need to be communicated
if endpoints were received successfully.