added NDK to docker and signing config test
for some reason, the NDK missing was a compile issue in the docker
compilation. fixed that.
the recent changes that were meant to facilitate release version
compilation with "wrong" signature to compare all but the signature with
the released APK made it difficult for the release manager to reliably
compile with the right signature. Added forced release signature via
local.properties