Remake tokens addition dialog so that eth account chosen first
Also removed subscription on AccountCreated event, otherwise we could leave
the activity when we still need to perform some things there, such as create
ERC20 account after Eth account. All the code from the subscription was
copied to places where it would be previously fired. Repetition is the flaw though.