renamed "Generic" classes
mbw was originally Bitcoin only, so a Transaction was a Bitcoin
transaction etc. but with the introduction of altcoins, for example
GenericTransaction was introduced to mean both EthTransaction and
(Bitcoin)Transaction. This is confusing, so with this commit, almost all
use of the term "Generic" is removed and all classes are being renamed
accordingly. I did in this order:
{,Bitcoin}Transaction
{Generic,}Transaction
{Generic,}TransactionSummary
{Generic,}TransactionData
{,Bitcoin}Address
{Generic,}Address
merged GenericModule into WalletModule
{Generic,}AssetInfo
{Generic,}AssetUri
{Generic,}AssetUriParser
{Generic,}Backing
{Generic,}BlockExplorerLabel
{Generic,}BuildTransactionException
Isufficient{Funds,Btc}Exception
{Generic,}InsufficientFundsException
{,Btc}OutputTooSmallException
{Generic,}OutputTooSmallException
{Generic,}TransactionBroadcastException
{Generic,}DetailsFragment
{Generic,}Fee
{Generic,}InputViewModel
{Generic,}OutputViewModel
BitcoinBased{Generic,}Transaction