avoid leaking KYC documents to shared folders
KYC documents get written into a private/protected, purpose-named folder
"tmp/bequant/Pictures". This avoids other apps from reading and
interference with other features of the same app that might also use a
"Pictures" folder.

The file is supposed to be deleted after use. By running delete() on it
on creation, it should get deleted after garbage collection of the last
reference to it.