fix convoluted inheritance in AccountContextImpl
AccountContextImpl worked like this:

open class Base<X: Base> (val listener: (X) -> Unit) {
    fun someFun() {
        ...
        listener(this as X)
    }
}

class Derived(listener: (Derived) -> Unit): Base<Derived>(listener)

This is problematic as `listener(this as X)` has to make assumptions that are not enforceable.

This change uses the conventional approach of an abstract function in a class that should have been abstract anyway.