Some fixes
CancelableRequest is an implementation detail of interruptSync and for BTC only so far.
cancel() may be null and should rather be null by default than throw an error.
cancel() being a var in a data object needs to be @Transient to not get serialized and sent to the server somehow.
Don't extend CancelableRequest if it's not being used.
null cancelableRequest after invoke()