Decouple scale in which fee estimations are returned by server and fee items building process
previously the code used for building fee items was using hardcoded
value of a scale with which fee estimation was returned by electrumx
server (i.e. in kbytes). since now this code is used for fee estimations
returned not only from electrumx we have to decouple this parts. scale
is now bound to FeeEstimations class