Fix crash bug by ensuring getTypeByName does not return null
If an unknown coin type is read from user data, it will cause getTypeByName to return null.  The currentCurrencyMap will then have an asset with key equal to null, which should never happen.  That will later cause a crash in CurrencySwitcher.setCurrency.  This fix here is to use the BTC coin type as the default instead of returning null.